feat: Base Layout for Error Page
